==> analysis/02_multilayer_landscape.R <==
#---
#title: "calculate multilayer landscape for performance params"
#author: "Eli"
#date: "2024-06-17"
#output: html_document
#editor_options:
#  chunk_output_type: console
#---

# import
library( here )
library( tidyverse )

# input
input <- read_tsv( file = here( 'results', 'simulated_bp_and_class.tsv.gz' ),
                   col_types = cols( .default = col_double(),
                                     Status = col_character( ) ) )

# explore
saved_data <- tibble()

for ( w1 in seq( from = 0, to = 150, by = 0.25 ) ) {
  for ( w2 in seq( from = 0, to = 150, by = 0.25 ) ) {
    current <- input %>%
      # adjust fxn (subtraction)
      mutate( in1 = SBP - w1 ) %>%
      mutate( in2 = DBP - w2 ) %>%
      # ReLU
      mutate( a1 = case_when(
        in1 > 0 ~ in1,
        TRUE ~ 0
      )) %>%
      mutate( a2 = case_when(
        in2 > 0 ~ in2,
        TRUE ~ 0
      )) %>%
      # heaviside
      mutate( heaviside = case_when(
        a1 > 0 ~ 1,
        a2 > 0 ~ 1,
        TRUE ~ 0 )) %>%
      # prediction
      mutate( prediction = case_when(
        heaviside > 0 ~ "HT",
        TRUE ~ "NT" )) %>%
      # classification stats
      mutate( isTP = case_when(
        Status == 'HT' & prediction == 'HT' ~ 1,
        TRUE ~ 0
      )) %>%
      mutate( isTN = case_when(
        Status == 'NT' & prediction == 'NT' ~ 1,
        TRUE ~ 0
      )) %>%
      mutate( isFP = case_when(
        Status == 'NT' & prediction == 'HT' ~ 1,
        TRUE ~ 0
      )) %>%
      mutate( isFN = case_when(
        Status == 'HT' & prediction == 'NT' ~ 1,
        TRUE ~ 0
      ))

    # classification stats
    TP = sum( current$isTP )
    FP = sum( current$isFP )
    TN = sum( current$isTN )
    FN = sum( current$isFN )

    current_precision = TP / ( TP + FP )
    current_recall = TP / ( TP + FN )

    if ( !is.finite( current_precision ) ) current_precision = 0.0
    if ( !is.finite( current_recall ) ) current_recall = 0.0

    # saving data
    saved_data <- tibble( w1 = w1,
                          w2 = w2,
                          tp = TP,
                          fp = FP,
                          tn = TN,
                          fn = FN,
                          precision = current_precision,
                          recall = current_recall ) %>%
      rbind( saved_data )
  }
}

write_tsv( x = saved_data,
           file = here( 'results', 'performance_landscape.tsv.gz' ) )

# session info
Sys.time()
getwd()
sessionInfo()
